+++ R/graph.R
@@ -71,6 +71,11 @@ graphplot <- function(x, measure = "support", shading = "lift",
   
   engines <- c("default", "ggplot2", "igraph", "interactive", "graphviz", 
     "visNetwork", "htmlwidget")
+  if(control$engine == "help") {
+    message("Available engines for this plotting method are:\n", paste0(engines, collapse = ", "))
+    return(invisible(engines))  
+  }
+  
   m <- pmatch(control$engine, engines, nomatch = 0)
   if(m == 0) stop("Unknown engine: ", sQuote(control$engine), 
     " Valid engines: ", paste(sQuote(engines), collapse = ", "))