+++ R/groupedmatrix.R
@@ -23,6 +23,11 @@ grouped_matrix_plot <- function(rules, measure, shading, control = NULL, ...){
   ## shading controls color
  
   engines <- c("default", "grid", "interactive", "ggplot2", "htmlwidget")
+  if(control$engine == "help") {
+    message("Available engines for this plotting method are:\n", paste0(engines, collapse = ", "))
+    return(invisible(engines))  
+  }
+  
   m <- pmatch(control$engine, engines, nomatch = 0)
   if(m == 0) stop("Unknown engine: ", sQuote(control$engine), 
     " Valid engines: ", paste(sQuote(engines), collapse = ", "))
@@ -170,6 +175,7 @@ rowMaxs <- function(x, na.rm=FALSE) apply(x, MARGIN=1, max, na.rm=na.rm)
   ma[!is.finite(ma)] <- NA
   ma
 }
+
 
 ## create an grouped_matrix
 grouped_matrix_int <- function(rules, measure, shading, control, plot = TRUE) {